Clean up datamodel tests folder structure
