+++ pkg/datamodel-tests/package.json
@@ -12,7 +12,7 @@
 		"lint": "prettier --check --plugin-search-dir=. . && eslint .",
 		"lint:strict": "eslint . --max-warnings=0",
 		"format": "prettier --write --plugin-search-dir=. .",
-		"test": "echo 'Running test runner unit tests' && vitest runner/__tests__",
+		"test": "echo 'Running test runner unit tests' && vitest __tests__",
 		"run-tests": "vitest main.test.ts",
 		"typecheck": "tsc --noEmit"
 	},
