+++ pkg/datamodel-tests/src/__testData__/books/science/0061983411.json
