+++ pkg/datamodel-tests/src/__testData__/books/science/0080524230.json
