+++ pkg/datamodel-tests/tsconfig.json
@@ -1,6 +1,14 @@
 {
 	"extends": "../scaffold/tsconfig.json",
 	"compilerOptions": {
+		"paths": {
+			"@/*": ["src/*"],
+			"@runner/*": ["src/runner/*"],
+			"@tests/*": ["src/tests/*"],
+			"@loaders/*": ["src/data-loaders/*"],
+			"@unit-tests/*": ["src/__tests__/*"],
+			"@unit-test-data/*": ["src/__testData__/*"]
+		},
 		"rootDir": "src",
 		"baseUrl": ".",
